refactor: extract login logic into separate file
